+++ usvg/src/convert/mod.rs
@@ -415,6 +415,11 @@ fn remove_empty_groups(tree: &mut tree::Tree) {
 }
 
 fn ungroup_groups(tree: &mut tree::Tree, opt: &Options) {
+    fn prepend_ts(ts1: &mut tree::Transform, mut ts2: tree::Transform) {
+        ts2.append(ts1);
+        *ts1 = ts2;
+    }
+
     fn ungroup(parent: tree::Node, opt: &Options) -> bool {
         let mut changed = false;
 
@@ -443,16 +448,16 @@ fn ungroup_groups(tree: &mut tree::Tree, opt: &Options) {
                     // Update transform.
                     match *child.borrow_mut() {
                         tree::NodeKind::Path(ref mut path) => {
-                            path.transform.append(&ts);
+                            prepend_ts(&mut path.transform, ts);
                         }
                         tree::NodeKind::Text(ref mut text) => {
-                            text.transform.append(&ts);
+                            prepend_ts(&mut text.transform, ts);
                         }
                         tree::NodeKind::Image(ref mut img) => {
-                            img.transform.append(&ts);
+                            prepend_ts(&mut img.transform, ts);
                         }
                         tree::NodeKind::Group(ref mut g) => {
-                            g.transform.append(&ts);
+                            prepend_ts(&mut g.transform, ts);
                         }
                         _ => {}
                     }